src/aws-lambda/process-image-upload/requirements.txt
pillow==8.0.1
pyzbar==0.1.8
psycopg2==2.8.6